@@ -8,7 +8,6 @@ use InvalidArgumentException as SplInvalidArgumentException;
 use Laminas\ServiceManager\AbstractFactoryInterface;
 use Laminas\ServiceManager\Initializer\InitializerInterface;
 
-use function get_class;
 use function gettype;
 use function is_object;
 use function sprintf;
@@ -18,29 +17,23 @@ use function sprintf;
  */
 class InvalidArgumentException extends SplInvalidArgumentException implements ExceptionInterface
 {
-    /**
-     * @param mixed $initializer
-     */
-    public static function fromInvalidInitializer($initializer): self
+    public static function fromInvalidInitializer(mixed $initializer): self
     {
         return new self(sprintf(
             'An invalid initializer was registered. Expected a callable or an'
             . ' instance of "%s"; received "%s"',
             InitializerInterface::class,
-            is_object($initializer) ? get_class($initializer) : gettype($initializer)
+            is_object($initializer) ? $initializer::class : gettype($initializer)
         ));
     }
 
-    /**
-     * @param mixed $abstractFactory
-     */
-    public static function fromInvalidAbstractFactory($abstractFactory): self
+    public static function fromInvalidAbstractFactory(mixed $abstractFactory): self
     {
         return new self(sprintf(
             'An invalid abstract factory was registered. Expected an instance of or a valid'
             . ' class name resolving to an implementation of "%s", but "%s" was received.',
             AbstractFactoryInterface::class,
-            is_object($abstractFactory) ? get_class($abstractFactory) : gettype($abstractFactory)
+            is_object($abstractFactory) ? $abstractFactory::class : gettype($abstractFactory)
         ));
     }
 }
