@@ -29,8 +29,7 @@ interface InitializerInterface extends Initializer\InitializerInterface
     /**
      * Initialize
      *
-     * @param mixed $instance
      * @return mixed
      */
-    public function initialize($instance, ServiceLocatorInterface $serviceLocator);
+    public function initialize(mixed $instance, ServiceLocatorInterface $serviceLocator);
 }
