@@ -68,15 +68,12 @@ EOH;
 
     private ConsoleHelper $helper;
 
-    private string $scriptName;
-
     /**
      * @param string $scriptName
      */
-    public function __construct($scriptName = self::DEFAULT_SCRIPT_NAME, ?ConsoleHelper $helper = null)
+    public function __construct(private $scriptName = self::DEFAULT_SCRIPT_NAME, ?ConsoleHelper $helper = null)
     {
-        $this->scriptName = $scriptName;
-        $this->helper     = $helper ?: new ConsoleHelper();
+        $this->helper = $helper ?: new ConsoleHelper();
     }
 
     /**
@@ -128,7 +125,6 @@ EOH;
     }
 
     /**
-     * @param array $args
      * @return object
      */
     private function parseArgs(array $args)
